Added support of AmeriaBank Business into bank_downloader.py
