+++ README.md
@@ -106,18 +106,20 @@ See list of supported banks, supported formats and relevant instructions in belo
   Supports all features native to app and Beancount reports.
   In `config.yaml` is referenced by `ameriaCsvFilesGlob` setting.
   Parsed by [ameria_csv_parser.go](/ameria_csv_parser.go).
+  Also supports files from https://business.myameria.am (new REST HTTP API based site).
+  While it doesn't provide all features and unstable it allows to download CSV statements semi-automatically.
   There is an option to **download CSV statements** via
-  [bank_downloader.py](/scripts/bank_downloader.py). You provide authentication and a **since** date; the script simulates opening each account and downloading CSV on its own (same behaviour as MyAmeria).
+  [bank_downloader.py](/scripts/bank_downloader.py). You provide a session cookie and a **since** date; the script uses the API to list accounts and download CSV for each.
   <details>
   <summary>How to download AmeriaBank Business statements</summary>
 
-  1. Log in at https://online.ameriabank.am (e.g. QR code via mobile app).
-  2. Open DevTools → **Network**. Find any request whose URL contains **Content.MainForm.wgx**.
-  3. Copy **Cookie** (full header) and **content_url** (full Request URL of that request).
-  4. In `bank_dowloader_config.yaml` set `ameriabank.cookie`, `ameriabank.content_url` and `ameriabank.since-DD-MM-YYYY` (same format as `my_ameria`). Optionally set `folder_path` (base folder for CSVs; empty = scripts folder) and `accounts` (list with `name`, `number`, `path` per account to limit and name files).
-  5. Run `python scripts/bank_downloader.py` (or `make bank-downloader`). The script opens Accounts, for each account opens Statement, sets date range (since → today), exports CSV and saves. With no `accounts` list, files are named `<account_number>_<account_name>_since_<since>.csv` under `folder_path`. You only supply cookie and content_url; the script obtains session tokens (LR) and export URLs (requestid) from the server responses — it does not ask you for them or guess them.
+  1. Log in at https://business.myameria.am. Need to provide OTP code which is located in mobile app. See Menu -> Settings -> OTP -> Copy Code.
+  2. Open DevTools → **Network**. Find any request to **gateway-businessmyameria.ameriabank.am**.
+  3. Copy the **Cookie** header (must include RefreshToken and TS* / AccessToken).
+  4. In `bank_dowloader_config.yaml` set `ameriabank.cookie` and `ameriabank.since-DD-MM-YYYY` (DD-MM-YYYY format). Optionally set `folder_path` (base folder for CSVs; empty = scripts folder) and `accounts` (list with `name`, `number`, `path` per account to limit and name files).
+  5. Run `python scripts/bank_downloader.py` (or `make bank-downloader`). The script calls the API to list settlement and card accounts, then downloads CSV for each from `since` until today. With no `accounts` list, files are named `<account_number>_<account_name>_since_<since>.csv` under `folder_path`. On 401 Unauthorized the script refreshes the cookie and retries.
 
-  Session (cookie) expires; re-copy cookie and content_url after re-login. Helpers in [scripts/bank_helpers.py](/scripts/bank_helpers.py).
+  Session (cookie) expires; re-copy cookie after re-login. Helpers in [scripts/bank_helpers_ameria.py](/scripts/bank_helpers_ameria.py).
   </details>
 - [NONE] AmeriaBank for Businesses XML (.xml) files downloaded per-account from
   https://online.ameriabank.am/InternetBank/MainForm.wgx
