+++ scripts/bank_dowloader_config.yaml.template
@@ -11,19 +11,17 @@ my_ameria:
   # How to name a resulting file. Should be matched by 'genericCsvFilesGlob' in config.yaml.
   history_path: "../generic MyAmeria History since 2024-09-01.csv"
 
-# AmeriaBank Business (business.myameria.am) for legal entities.
-# Cookie: copy from DevTools → Network → any request to gateway-businessmyameria.ameriabank.am.
-#   Must include RefreshToken and TS0180e8bb (or similar TS*). AccessToken can be expired —
-#   on 401 we call Authentication/Refresh and retry automatically.
-# Parameters:
-#   cookie    — full Cookie header (TS*=...; RefreshToken=...; AccessToken=...)
-#   since-DD-MM-YYYY — start date for statements; end date is always today.
+# AmeriaBank (business.myameria.am) – HTTP REST API only.
 ameriabank:
+  # Log in at https://business.myameria.am, open DevTools → Network, copy Cookie header value from any request to gateway-businessmyameria.ameriabank.am.
+  # Cookie must include RefreshToken and TS* (AccessToken);
   cookie: "TS0180e8bb=...; RefreshToken=...; AccessToken=..."
+  # Start day to get statement. Always until today.
   since-DD-MM-YYYY: "01-09-2024"
+  # Where to put resulting files. There would be one file per account with name "AccountStatement {acc.number} {acc.name} since {since-DD-MM-YYYY}.csv".
   folder_path: ""
-  # Optional: list accounts to download (by number or name); path overrides default filename.
+  # Optional: list only specific accounts to download (by number or name); path overrides "folder_path" above.
   # accounts:
   #   - name: "Settlement AMD"
-  #     number: "1570048149171400"
+  #     number: "15700481491234567"
   #     path: "../my_settlement_amd.csv"
